Project management, NW.js deployment support
- Change name of app.js to server.js
- Add index.html to root to support NW.js deployment
- Set version to 1.0.0-ALPHA
- Remove junk from style.css
- Forgot to commit changes for api.js
- Rename close method in db.js
- Make small changes to toby-ui.js to support setting window title from
a socket.io message.
- Add method to youtube-ui.js to support snapping back to the YouTube
player after exiting fullscreen in NW.js since it snaps to top of page
(bug in NW.js?)
- Make socket a global and put it in a script tag in index.hbs template
so that all React components have access to the socket.
- Remove the NW.js specific code from the layout.hbs template because
that is now contained in the root level index.html file which is
specific and there to support deployment in NW.js